Remove PriceList class
