+++ src/RehauSku.Assist.csproj
@@ -124,9 +124,8 @@
     <Compile Include="PriceListTools\CombineTool.cs" />
     <Compile Include="PriceListTools\AbstractPriceListTool.cs" />
     <Compile Include="PriceListTools\MergeTool.cs" />
-    <Compile Include="PriceListTools\PriceList.cs" />
     <Compile Include="PriceListTools\PriceListPosition.cs" />
-    <Compile Include="PriceListTools\PriceListSheet.cs" />
+    <Compile Include="PriceListTools\PriceList.cs" />
     <Compile Include="Ribbon\RibbonController.cs" />
     <Compile Include="Assistant\HttpClientUtil.cs" />
     <Compile Include="Assistant\StoreResponse.cs" />
